% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_cols.R
\name{add_idx_y_col}
\alias{add_idx_y_col}
\title{Add \code{idx_y} column}
\usage{
add_idx_y_col(data)
}
\arguments{
\item{data}{Tracking data with \code{y_side} column}
}
\description{
Add \code{idx_y} column describing receiver's position relative
to the sideline on one side of the field. \code{y_side} column is expected
to already exist in \code{data}
(so \code{add_y_side_col} should have already been called).
}
