% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_feature_engineering.R
\name{add_receiver_target_rates}
\alias{add_receiver_target_rates}
\title{add_receiver_target_rates calculate how frequently receivers are targeted in entire dataset}
\usage{
add_receiver_target_rates(df)
}
\arguments{
\item{df}{DataFrame of target data}
}
\value{
DataFrame with the target rates added on
}
\description{
add_receiver_target_rates calculate how frequently receivers are targeted in entire dataset
}
