% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_cols.R
\name{add_side_cols}
\alias{add_side_cols}
\title{Add a \code{x_side} and \code{y_side} columns}
\usage{
add_side_cols(data, cutoff_x = 2.5, cutoff_y = 3)
}
\arguments{
\item{data}{Tracking data}

\item{cutoff_x}{\code{cutoff} for \code{add_x_side()}}

\item{cutoff_y}{\code{cutoff} for \code{add_y_side()}}
}
\description{
Add a \code{x_side} and \code{y_side} columns
}
\seealso{
\code{\link{add_x_side}} \code{\link{add_y_side}}
}
