% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering_throw.R
\name{add_throw_vector_to_positions}
\alias{add_throw_vector_to_positions}
\title{add_throw_vector_to_positions takes player positions at release along with throw information to calculate how close the player is to the throw}
\usage{
add_throw_vector_to_positions(release_positions, throw_vectors)
}
\arguments{
\item{release_positions}{player positions at release of ball}

\item{throw_vectors}{information about the pass}
}
\value{
a dataframe with the distance & time the player has to reach the throw
}
\description{
add_throw_vector_to_positions takes player positions at release along with throw information to calculate how close the player is to the throw
}
