% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_cols.R
\name{add_x_side_col}
\alias{add_x_side_col}
\title{Add \code{x_side} column}
\usage{
add_x_side_col(data, cutoff = 2.5)
}
\arguments{
\item{data}{Tracking data}

\item{cutoff}{Yards behind ball (towards backfield) within which to consider \code{"los"} (line of scrimmage).}
}
\description{
If player's \code{x} is less than \code{abs(ball_x - cutoff)}
(accounting for \code{play_direction}), then \code{x_side} is
\code{"los"} (line of scrimmage).
Otherwise, it is \code{"backfield"}.
}
\examples{
\dontrun{
library(bdb2021)
library(tidyverse)
tracking <- import_tracking(1)
snap_frames <- tracking \%>\% filter(event == 'ball_snap')
# Used this to determine `cutoff` for `x_side`
snap_frames \%>\%
  filter(position == 'RB') \%>\%
  inner_join(plays) \%>\%
  mutate(x_relative = x - ball_x) \%>\%
  ggplot() +
  aes(x = x_relative) +
  geom_histogram(binwidth = 1) +
  facet_grid(offense_formation~play_direction)
}
}
