% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_cols.R
\name{add_y_side_col}
\alias{add_y_side_col}
\title{Add \code{y_side} column}
\usage{
add_y_side_col(data, cutoff = 3)
}
\arguments{
\item{data}{Tracking data}

\item{cutoff}{Yards horizontally along same of plane of ball to consider \code{"mid"}.}
}
\description{
If player's \code{y} is less than or more than \code{abs(ball_y - cutoff)}
(accounting for \code{play_direction}), then \code{y_side} is
\code{"left"} or \code{"right"}.
Otherwise, it is \code{"mid"}.
}
\examples{
\dontrun{
library(bdb2021)
library(tidyverse)
tracking <- import_tracking(1)
snap_frames <- tracking \%>\% filter(event == 'ball_snap')
# Used this to determine `cutoff` for `y_side`
snap_frames \%>\%
  filter(position == 'TE') \%>\%
  inner_join(plays) \%>\%
  mutate(
    y_relative = y - ball_y,
    across(y_relative, ~cut(.x, breaks = seq(-8, 8, by = 0.5)))
  ) \%>\%
  ggplot() +
  aes(x = y_relative) +
  geom_bar() +
  facet_grid(offense_formation~.)
}
}
