% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{aggregate_week_files}
\alias{aggregate_week_files}
\title{aggregate_week_files aggregate all of the PBP data}
\usage{
aggregate_week_files()
}
\value{
a data.frame with the PBP data for the specified week
}
\description{
aggregate_week_files aggregate all of the PBP data
}
