% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary_fct_to_lgl}
\alias{binary_fct_to_lgl}
\title{Binary 1 and 0 to TRUE and FALSE}
\usage{
binary_fct_to_lgl(x)
}
\arguments{
\item{x}{factor}
}
\description{
Convert numerics to logicals, mostly for modeling purposes.
}
