% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_raw_stat_table.R
\name{build_raw_stat_table}
\alias{build_raw_stat_table}
\title{build_raw_stat_table Get data on speed, acceleration to use in shiny app}
\usage{
build_raw_stat_table()
}
\description{
build_raw_stat_table Get data on speed, acceleration to use in shiny app
}
