% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_target_data.R
\name{build_target_results}
\alias{build_target_results}
\title{build_target_results output all of the results of the target models to be used}
\usage{
build_target_results(final_model, scale_model, prior_target_model, data)
}
\arguments{
\item{final_model}{final result of the target model}

\item{scale_model}{logit model used to calibrate final results}

\item{prior_target_model}{model for target probability before play begins}

\item{data}{dataframe with all observations from season}
}
\value{
a success string
}
\description{
build_target_results output all of the results of the target models to be used
}
