% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering.R
\name{build_weather_df}
\alias{build_weather_df}
\title{build_weather_df returns a dataframe with parsed weather information for each gameId}
\usage{
build_weather_df(weeks_to_use)
}
\arguments{
\item{weeks_to_use}{weeks to get weather data for}
}
\value{
a data frame of forecasts matched to the appropriate game
}
\description{
build_weather_df returns a dataframe with parsed weather information for each gameId
}
