% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{catch_prob_diagnostic_plots}
\alias{catch_prob_diagnostic_plots}
\title{recalc_prob recalculate the catch probabilities after removing a defender}
\usage{
catch_prob_diagnostic_plots(train, test, xgb_model, logit_model)
}
\arguments{
\item{train}{the training set}

\item{test}{the test set}

\item{xgb_model}{the xgboost model}

\item{logit_model}{the logit model used for Platt scaling}
}
\value{
A success string
}
\description{
recalc_prob recalculate the catch probabilities after removing a defender
}
