% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_distance.R
\name{compute_min_distances}
\alias{compute_min_distances}
\title{Compute min distances between players.}
\usage{
compute_min_distances(o, d)
}
\arguments{
\item{data}{A data.frame with 4 columns: \code{nfl_id}, \code{side} (either "O" or "D"), \code{x}, \code{y}.}
}
\value{
A tibble with 6 columns: \code{nfl_id_o}, \code{nfl_id_d}, \code{x_o}, \code{x_d}, \code{y_o}, \code{y_d}.
}
\description{
Compute "optimal" nearest assignments of defensive players to individual offensive and defensive players (not including QB) in a single frame using the Hungarian method.
}
