% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_heroku_postgres.R
\name{connect_to_heroku_postgres}
\alias{connect_to_heroku_postgres}
\title{connect_to_heroku_postgres}
\usage{
connect_to_heroku_postgres()
}
\value{
the connection
}
\description{
connect to our Heroku Postgres instance
}
