% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_encrypted.R
\name{convert_secret}
\alias{convert_secret}
\title{convert_secret}
\usage{
convert_secret(secret)
}
\arguments{
\item{secret}{the string secret}
}
\value{
the raw secret
}
\description{
A function to convert the secret from char to raw
}
