% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering.R
\name{create_throw_vectors}
\alias{create_throw_vectors}
\title{create_throw_vectors returns a data frame with throw vectors}
\usage{
create_throw_vectors(football_data, throw_midpoint_frame_id)
}
\arguments{
\item{football_data}{a dataframe of football position data}

\item{throw_midpoint_frame_id}{dataframe with the midpoint of the throw for each play}
}
\value{
a data frame with information about the throw vector
}
\description{
create_throw_vectors returns a data frame with throw vectors
}
