% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divvy_catch_credit.R
\name{divvy_credit}
\alias{divvy_credit}
\title{divvy_credit divvy credit between defenders}
\usage{
divvy_credit(data, xgb_model, logit_model)
}
\arguments{
\item{data}{the data frame from which the probabilities are being predicted}

\item{xgb_model}{the xgboost model}

\item{logit_model}{the logit model used for Platt scaling}
}
\value{
a data frame of the weights to assign credit based on
}
\description{
divvy_credit divvy credit between defenders
}
