% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering_throw.R
\name{do_catch_prob_throw_feat_eng}
\alias{do_catch_prob_throw_feat_eng}
\title{do_catch_prob_throw_feat_eng build a data frame of catch prob features}
\usage{
do_catch_prob_throw_feat_eng(weeks_to_use = 1:17)
}
\arguments{
\item{weeks_to_use}{Numeric: a numeric vector of weeks to use (default 1:17)}
}
\value{
a data frame of features
}
\description{
do_catch_prob_throw_feat_eng build a data frame of catch prob features
}
