% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_feature_engineering.R
\name{do_target_prob_feature_eng}
\alias{do_target_prob_feature_eng}
\title{do_target_prob_feat_eng build a data frame of target prob features}
\usage{
do_target_prob_feature_eng(weeks_to_use = 1:17)
}
\arguments{
\item{weeks_to_use}{Numeric: a numeric vector of weeks to use (default 1:17)}
}
\value{
a data frame of features
}
\description{
do_target_prob_feat_eng build a data frame of target prob features
}
