% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_play.R
\name{.filter_side}
\alias{.filter_side}
\title{Filter by team}
\usage{
.filter_side(data, side = c("o", "d"))
}
\description{
Filter to one side of the ball (offense or defense).
}
\details{
This should only be used if you know what you're doing.
}
