% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_play.R
\name{.fix_d}
\alias{.fix_d}
\title{Ignore rushers}
\usage{
.fix_d(o, d, rushers)
}
\arguments{
\item{o, d}{Offense and defense matrices}

\item{rushers}{matrix for rushers}
}
\description{
Ignore rushers on defensive side of ball (in \code{d}) in min distance calculation.
}
\details{
This should only be used if you know what you're doing.
}
