% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.get_dir_in}
\alias{.get_dir_in}
\title{Get input directory}
\usage{
.get_dir_in()
}
\description{
Get directory where file is stored, which is dependent on whether the code is run locally or on Kaggle.
Best to change with \verb{options(bdb2021.dir_in = <dir>)} so that other \verb{import_*()}
}
