% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{.round_any}
\alias{.round_any}
\title{Round to multiple of any number.}
\usage{
.round_any(x, accuracy, f = round)
}
\description{
Round to multiple of any number.
}
\details{
Basically like \code{\link[plyr]{round_any}}
}
\seealso{
\url{http://search.r-project.org/library/plyr/html/round_any.html}
}
