% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_catch_prob_xgb}
\alias{fit_catch_prob_xgb}
\title{fit_catch_prob_xgb fit the xgboost catch prob model}
\usage{
fit_catch_prob_xgb(workflow, pars, data_split, data)
}
\arguments{
\item{workflow}{a workflow object}

\item{pars}{a tibble of the parameters (generally output from tune::select_best())}

\item{data_split}{a data_split object}

\item{data}{a full data frame to fit the model on}
}
\value{
a list with the final xgboost model and the modeling results (auc, acc, etc.)
}
\description{
fit_catch_prob_xgb fit the xgboost catch prob model
}
