% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_logit_platt_scaler}
\alias{fit_logit_platt_scaler}
\title{fit_logit_platt_scaler fit the Platt scaler to calibrate the xgboost predictions}
\usage{
fit_logit_platt_scaler(model, data)
}
\arguments{
\item{model}{an xgboost model fit with parsnip}

\item{data}{a data frame}
}
\value{
a logistic regression (parsnip) model
}
\description{
fit_logit_platt_scaler fit the Platt scaler to calibrate the xgboost predictions
}
