% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_prior_target_prob}
\alias{fit_prior_target_prob}
\title{fit_prior_target_prob create an estimate of the likelihood a receiver is targeted before the play starts}
\usage{
fit_prior_target_prob(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a logistic regression model of pre-play target probabilities
}
\description{
fit_prior_target_prob create an estimate of the likelihood a receiver is targeted before the play starts
}
