% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_target_prob_rf}
\alias{fit_target_prob_rf}
\title{fit_target_prob_rf fit the rf target prob model}
\usage{
fit_target_prob_rf(workflow, pars, data_split, data)
}
\arguments{
\item{workflow}{a workflow object}

\item{pars}{a tibble of the parameters (generally output from tune::select_best())}

\item{data_split}{a data_split object}

\item{data}{a full data frame to fit the model on}
}
\value{
a list with the final random forest model and the modeling results (auc, acc, etc.)
}
\description{
fit_target_prob_rf fit the rf target prob model
}
