% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_prob.R
\name{fit_target_prob_split}
\alias{fit_target_prob_split}
\title{Fit target probabilities}
\usage{
fit_target_prob_split(
  trn,
  tst,
  fmla,
  min_n = 2,
  mtry = 42,
  trees = 500,
  suffix = "",
  overwrite = FALSE
)
}
\description{
For use with big boy data
}
