% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{generate_formula}
\alias{generate_formula}
\title{Generate a model formula}
\usage{
generate_formula(
  data,
  y,
  x = NULL,
  x_include = NULL,
  x_exclude = NULL,
  rgx_x_include = NULL,
  rgx_x_exclude = NULL,
  intercept = TRUE
)
}
\arguments{
\item{data}{data.frame}

\item{y}{character representing the response variable in the \code{data}}

\item{x}{character vector representing the independent variables to include in the resulting formula}

\item{x_include, x_exclude}{explicit variables (or columns) in \code{data} to include or not include in the model formula}

\item{rgx_x_include, rgx_x_exclude}{Like \code{x_include} and \code{x_exclude}, but regular expressions}

\item{intercept}{Boolean indicating whether to include an intercept term}
}
\description{
Generate a model formula
}
