% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man_zone_class.R
\name{generate_man_zone_class}
\alias{generate_man_zone_class}
\title{full feature engineering and classification for man v zone coverages
WARNING: this takes a lot of time to run}
\usage{
generate_man_zone_class()
}
\value{
a dataframe of man vs zone predictions for non-lineman on a given play
}
\description{
full feature engineering and classification for man v zone coverages
WARNING: this takes a lot of time to run
}
