% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_constants}
\alias{get_constants}
\title{get_constants get project-wide constants}
\usage{
get_constants(const)
}
\arguments{
\item{const}{the constant to get (one of defense_positions, offense_positions, throw_start_events, throw_end_events, defensive_back_positions, linebacker_positions, and defensive_line_positions)}
}
\value{
string: the constant
}
\description{
get_constants get project-wide constants
}
