% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering_arrival.R
\name{get_defense_locs_at_arrival}
\alias{get_defense_locs_at_arrival}
\title{get_defense_locs_at_arrival returns a data from of the locations of the defense at throw time}
\usage{
get_defense_locs_at_arrival(player_locs_at_arrival)
}
\arguments{
\item{player_locs_at_arrival}{a data frame of player locations at throw time}
}
\value{
a data frame with the locations of the defenders at throw time
}
\description{
get_defense_locs_at_arrival returns a data from of the locations of the defense at throw time
}
