% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering_throw.R
\name{get_defense_locs_at_throw}
\alias{get_defense_locs_at_throw}
\title{get_defense_locs_at_throw returns a data from of the locations of the defense at throw time}
\usage{
get_defense_locs_at_throw(player_locs_at_throw, throw_vectors)
}
\arguments{
\item{player_locs_at_throw}{a data frame of player locations at throw time}

\item{throw_vectors}{the throw vectors}
}
\value{
a data frame with the locations of the defenders at throw time
}
\description{
get_defense_locs_at_throw returns a data from of the locations of the defense at throw time
}
