% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering.R
\name{get_football_location_at_arrival}
\alias{get_football_location_at_arrival}
\title{get_football_location_at_arrival returns a data frame of the location of the football on each play}
\usage{
get_football_location_at_arrival(pbp)
}
\arguments{
\item{pbp}{a dataframe of play-by-play data}
}
\value{
a data frame with the positions of the football at arrival
}
\description{
get_football_location_at_arrival returns a data frame of the location of the football on each play
}
