% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering.R
\name{get_football_locations}
\alias{get_football_locations}
\title{get_football_locations get a data frame of the location of the football on each play}
\usage{
get_football_locations(pbp)
}
\arguments{
\item{pbp}{a dataframe of play-by-play data}
}
\value{
a data frame with the target data
}
\description{
get_football_locations get a data frame of the location of the football on each play
}
