% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering.R
\name{get_heights}
\alias{get_heights}
\title{get_heights returns a data frame of player heights}
\usage{
get_heights()
}
\value{
a data frame of the heights
}
\description{
get_heights returns a data frame of player heights
}
