% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_intersection_cutoff}
\alias{get_intersection_cutoff}
\title{Pick route cutoff}
\usage{
get_intersection_cutoff()
}
\description{
Threshold (in seconds) beyond which receiver intersections are no longer consider picks
}
