% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering.R
\name{get_max_throw_velo}
\alias{get_max_throw_velo}
\title{get_max_throw_velo returns a data frame of maximum throw velos for each play}
\usage{
get_max_throw_velo(pbp)
}
\arguments{
\item{pbp}{a dataframe of play-by-play data}
}
\value{
a data frame with the game id, the play id, and the max throw velo
}
\description{
get_max_throw_velo returns a data frame of maximum throw velos for each play
}
