% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering.R
\name{get_pi_and_sack}
\alias{get_pi_and_sack}
\title{get_pi_and_sack returns a data frame of pass interferences and sacks}
\usage{
get_pi_and_sack(plays)
}
\arguments{
\item{plays}{a dataframe of plays}
}
\value{
a data frame with the positions of the football at arrival
}
\description{
get_pi_and_sack returns a data frame of pass interferences and sacks
}
