% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering.R
\name{get_play_outcomes}
\alias{get_play_outcomes}
\title{get_play_outcomes returns a data frame of play outcomes}
\usage{
get_play_outcomes(pbp, pi_or_sack)
}
\arguments{
\item{pbp}{the play by play data}

\item{pi_or_sack}{the play result data}
}
\value{
a data frame with the play outcomes
}
\description{
get_play_outcomes returns a data frame of play outcomes
}
