% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering_arrival.R
\name{get_target_location_at_arrival}
\alias{get_target_location_at_arrival}
\title{get_target_location_at_arrival returns a data frame of the locations of the targeted receiver at throw time on each play}
\usage{
get_target_location_at_arrival(pbp, football_data, targeted_receiver)
}
\arguments{
\item{pbp}{a dataframe of play-by-play data}

\item{football_data}{the football data}

\item{targeted_receiver}{the targets data}
}
\value{
a data frame with the positions
}
\description{
get_target_location_at_arrival returns a data frame of the locations of the targeted receiver at throw time on each play
}
