% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering_throw.R
\name{get_targeted_receiver}
\alias{get_targeted_receiver}
\title{get_targeted_receiver get a data frame of the targeted receiver on each play}
\usage{
get_targeted_receiver()
}
\value{
a data frame with the target data
}
\description{
get_targeted_receiver get a data frame of the targeted receiver on each play
}
