% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_prob_feature_engineering.R
\name{get_throw_dists}
\alias{get_throw_dists}
\title{get_throw_dists returns a data frame of throw distances on each play}
\usage{
get_throw_dists(pbp, throw_start_events, throw_end_events)
}
\arguments{
\item{pbp}{a dataframe of play-by-play data}

\item{throw_start_events}{a character vector of throw start events}

\item{throw_end_events}{a character vector of throw end events}
}
\value{
a data frame with the game id, the play id, and the throw distance
}
\description{
get_throw_dists returns a data frame of throw distances on each play
}
