% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{get_weather}
\alias{get_weather}
\title{get_weather return the weather forecast in a given week}
\usage{
get_weather(week, year)
}
\arguments{
\item{week}{numeric, the week to get}

\item{year}{numeric, the year to get}
}
\value{
a data frame of features
}
\description{
get_weather return the weather forecast in a given week
}
