% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_field.R
\name{gg_field}
\alias{gg_field}
\title{Plot a field}
\usage{
gg_field(
  yardmin = 0,
  yardmax = 120,
  buffer = 5,
  direction = "horiz",
  field_color = "white",
  line_color = "black",
  sideline_color = field_color,
  endzone_color = field_color
)
}
\arguments{
\item{yardmin, yardmax, buffer}{numbers}

\item{direction}{\code{"horiz"} or \code{"vert"}}

\item{field_color}{color}

\item{line_color}{color}

\item{sideline_color}{color}

\item{endzone_color}{color}
}
\description{
Plot a field
}
\seealso{
\url{https://github.com/mlfurman3/gg_field}
}
