% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_tables.R
\name{gt_theme_538}
\alias{gt_theme_538}
\title{gt_theme_538 the Fivethirtyeight GT theme}
\usage{
gt_theme_538(data, ...)
}
\arguments{
\item{data}{the data to use to build the table}

\item{...}{other parameters}
}
\value{
list: the table in data frame form and gt form
}
\description{
gt_theme_538 the Fivethirtyeight GT theme
}
