% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heroku_postgres_prep.R
\name{heroku_postgres_inserts}
\alias{heroku_postgres_inserts}
\title{heroku_postgres_inserts}
\usage{
heroku_postgres_inserts()
}
\value{
the connection
}
\description{
connect to our Heroku Postgres instance
}
