% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{identify_intersection}
\alias{identify_intersection}
\title{Identify route intersection}
\usage{
identify_intersection(data, pad_backwards = TRUE, pad_x = 1)
}
\arguments{
\item{data}{Tracking data for a single play.}

\item{pad_backwards}{Boolean for whether to extend route backwards.}

\item{pad_x}{If \code{pad_backwards = TRUE}, this is the number of yards to extend the route.}
}
\description{
Identify if two routes intersect.
}
\examples{
\dontrun{
identify_intersection()
}
}
