% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_colors}
\alias{import_colors}
\title{Import team colors data}
\usage{
import_colors()
}
\description{
Import team colors data
}
\examples{
\dontrun{
import_colors()
}
}
\seealso{
\url{https://github.com/leesharpe/nfldata/blob/master/data/teamcolors.csv}
}
