% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_epa_model_bdb}
\alias{import_epa_model_bdb}
\title{Import package-generated EPA model based on \code{{nflfastR}}'s model}
\usage{
import_epa_model_bdb()
}
\description{
Import package-generated EPA model based on \code{{nflfastR}}'s model
}
