% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_min_dists_naive_target}
\alias{import_min_dists_naive_target}
\title{Import package-generated half-second min distances to target receiver}
\usage{
import_min_dists_naive_target()
}
\description{
Import package-generated half-second min distances to target receiver
}
