% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_new_features}
\alias{import_new_features}
\title{Import package-generated half-second features}
\usage{
import_new_features()
}
\description{
The \code{new_} part is to indicate that this set of features was created with
re-laxed criteria for plays. Specifically, more or less than five receivers
are allowed for play. (Previously, plays with more or less receivers
were thrown out.)
}
