% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_raw_target_probs}
\alias{import_raw_target_probs}
\title{Import package-generated raw half-second target probabilities}
\usage{
import_raw_target_probs()
}
\description{
Import package-generated raw half-second target probabilities
}
