% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_target_prob_model}
\alias{import_target_prob_model}
\title{Import package-generated clean target probability model}
\usage{
import_target_prob_model()
}
\description{
Import package-generated clean target probability model
}
