% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_week}
\alias{import_week}
\title{Import tracking data for a given week}
\usage{
import_week(
  week,
  positions = import_positions(),
  standardize = TRUE,
  dir = .get_dir()
)
}
\arguments{
\item{week}{Number between 1 and 17}

\item{positions}{Positions data for adding \code{side} column}

\item{standardize}{Boolean. Whether to \code{x} and \code{y} based on line of scrimmage (\code{los}) and \code{play_direction}. Note that \code{los} column
is added if TRUE.}
}
\description{
In addition to importing tracking data, moves \code{"Football"}
rows to columns \code{ball_x} and \code{ball_y} and adds \code{side}
column using \code{positions} data. If \code{standardize = TRUE},
then \code{los} column is also added.
}
\examples{
\dontrun{
import_week(1)
}
}
