% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_encrypted.R
\name{load_encrypted}
\alias{load_encrypted}
\title{load_encrypted}
\usage{
load_encrypted(file = "")
}
\arguments{
\item{file}{the file to load}
}
\value{
the decrypted file
}
\description{
A function to load encrypted .Rdata files
}
