% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_tables.R
\name{load_player_summary_table}
\alias{load_player_summary_table}
\title{load_player_summary_table Function that builds the overall results}
\usage{
load_player_summary_table()
}
\description{
load_player_summary_table Function that builds the overall results
}
