% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_tables.R
\name{make_catch_prob_table}
\alias{make_catch_prob_table}
\title{make_catch_prob_table makes the drops added table}
\usage{
make_catch_prob_table(
  arrival_data,
  throw_data,
  num = 1000,
  playcutoff = 300,
  show_top = TRUE
)
}
\arguments{
\item{num}{the number of players to include in the table (default: 1000)}

\item{playcutoff}{the min number of plays (default: 300)}

\item{show_top}{True sorts by the top players, False sorts by the worst}
}
\value{
list: the table in data frame form and gt form
}
\description{
make_catch_prob_table makes the drops added table
}
