% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_tables.R
\name{make_tendency_table}
\alias{make_tendency_table}
\title{make_tendency_table makes the deterrence table}
\usage{
make_tendency_table(data, model)
}
\arguments{
\item{data}{the data to use to build the table}

\item{xgb_model}{the xgboost model}

\item{logit_model}{the logit model for platt scaling}

\item{num}{the number of players to include in the table (default: 1000)}

\item{playcutoff}{the min number of plays (default: 300)}

\item{show_top}{True sorts by the top players, False sorts by the worst}
}
\value{
list: the table in data frame form and gt form
}
\description{
make_tendency_table makes the deterrence table
}
