% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_from_data}
\alias{read_from_data}
\title{read_from_data load data from csv file}
\usage{
read_from_data(file)
}
\arguments{
\item{file}{the file to load}
}
\value{
data.frame with the loaded file
}
\description{
read_from_data load data from csv file
}
