% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_individual_week}
\alias{read_individual_week}
\title{read_individual_week load a single week of PBP data}
\usage{
read_individual_week(week)
}
\value{
a data.frame with the PBP data for the specified week
}
\description{
read_individual_week load a single week of PBP data
}
