% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_non_week_files}
\alias{read_non_week_files}
\title{read_non_week_files read all of the non-PBP data}
\usage{
read_non_week_files()
}
\value{
a list of data.frames with the loaded files
}
\description{
read_non_week_files read all of the non-PBP data
}
