% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_targets}
\alias{read_targets}
\title{read_targets load targeted receiver data}
\usage{
read_targets()
}
\value{
a data.frame with the targets data
}
\description{
read_targets load targeted receiver data
}
