% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{round_any}
\alias{round_any}
\title{Round to multiple of any number.}
\source{
https://stackoverflow.com/questions/43627679/round-any-equivalent-for-dplyr/46489816#46489816
}
\usage{
round_any(x, accuracy, f = round)
}
\description{
Round to multiple of any number.
}
\details{
\code{\link[plyr]{round_any}}
}
\seealso{
\url(http://search.r-project.org/library/plyr/html/round_any.html)
}
