% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{run_target_prob_tuning_pipeline}
\alias{run_target_prob_tuning_pipeline}
\title{run_target_prob_tuning_pipeline pipeline for target prob model}
\usage{
run_target_prob_tuning_pipeline(tune_file_name)
}
\arguments{
\item{tune_file_name}{if the pipeline has been run before, the location of the tuning file}
}
\description{
run_target_prob_tuning_pipeline pipeline for target prob model
}
