% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_play.R
\name{save_animation}
\alias{save_animation}
\title{Save animation}
\usage{
save_animation(
  anim,
  height = 600,
  width = width,
  fps = 10,
  end_pause = fps,
  file = deparse(substitute(anim)),
  ext = "gif",
  dir = get_bdb_dir_figs(),
  path = file.path(dir, sprintf("\%s.\%s", file, ext)),
  renderer = gganimate::gifski_renderer(path),
  ...
)
}
\arguments{
\item{anim}{animation object}
}
\description{
Save animation
}
