% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_encrypted.R
\name{save_encrypted}
\alias{save_encrypted}
\title{save_encrypted}
\usage{
save_encrypted(data, file = "")
}
\arguments{
\item{data}{the object to encrypt}

\item{file}{the file path to save to}
}
\value{
NULL (invisible)
}
\description{
A function to encrypt and save Rdata files
}
