% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_play.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot}
\usage{
save_plot(
  object,
  dir = get_bdb_dir_figs(),
  file = deparse(substitute(object)),
  ext = "png",
  path = file.path(dir, sprintf("\%s.\%s", file, ext)),
  height = 8,
  width = height,
  ...
)
}
\arguments{
\item{object}{Plot/animation to save}

\item{dir}{Directory, file name (without extension), and extension used to generate \code{path} if \code{path} is not explicitly provided.}

\item{file}{Directory, file name (without extension), and extension used to generate \code{path} if \code{path} is not explicitly provided.}

\item{ext}{Directory, file name (without extension), and extension used to generate \code{path} if \code{path} is not explicitly provided.}

\item{path}{Path where to save plot/animation.}

\item{height, width}{Height and width to use for saved image, in inches.}

\item{...}{Passed to \code{gganimate::animate()}.}
}
\description{
Save a plot
}
