% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_position_coordinates}
\alias{standardize_position_coordinates}
\title{standardize_position_coordinates Takes raw position data and normalizes it so that drives only go one way and X coordinates are relative to goal line rather than back of end zone}
\usage{
standardize_position_coordinates(df)
}
\arguments{
\item{df}{DataFrame of position data to be transformed}
}
\value{
data.frame: normalized positional data
}
\description{
standardize_position_coordinates Takes raw position data and normalizes it so that drives only go one way and X coordinates are relative to goal line rather than back of end zone
}
