% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{target_prob_diagnositc_plots}
\alias{target_prob_diagnositc_plots}
\title{target_prob_diagnostic_plots same diagnostic plots for target probs}
\usage{
target_prob_diagnositc_plots(train, test, xgb_model, logit_model)
}
\arguments{
\item{train}{dataframe of training set}

\item{test}{dataframe of testing set}

\item{xgb_model}{xgboost model produced by tune_target_prob_xgb}

\item{logit_model}{model used for Pratt scaling}
}
\value{
success string
}
\description{
target_prob_diagnostic_plots same diagnostic plots for target probs
}
