% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_models.R
\name{tune_catch_prob_xgb}
\alias{tune_catch_prob_xgb}
\title{tune_catch_prob_xgb tune the xgboost catch prob model}
\usage{
tune_catch_prob_xgb(data, mod = "")
}
\arguments{
\item{data}{a full data frame to tune the model on}

\item{mod}{'a' for arrival, 't' for throw}
}
\value{
a list with the data, the data_split, the workflow, the best set of tuning parameters, and the tuning results
}
\description{
tune_catch_prob_xgb tune the xgboost catch prob model
}
