% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_models.R
\name{tune_target_prob_rf}
\alias{tune_target_prob_rf}
\title{tune_target_prob_rf tune the rf target prob model}
\usage{
tune_target_prob_rf(data, ncores = 4)
}
\arguments{
\item{data}{a full data frame to tune the model on}

\item{ncores}{the number of cores to use when tuning, defaults to 4}
}
\value{
a list with the data, the data_split, the workflow, the best set of tuning parameters, and the tuning results
}
\description{
tune_target_prob_rf tune the rf target prob model
}
