% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_models.R
\name{tune_target_prob_xgb}
\alias{tune_target_prob_xgb}
\title{tune_target_prob_xgb tune the XGB target probability model
TODO: actually document this rather than just copy paste}
\usage{
tune_target_prob_xgb(data)
}
\arguments{
\item{data}{a full data frame to tune the model on}
}
\value{
a list with the data, the data_split, the workflow, the best set of tuning parameters, and the tuning results
}
\description{
tune_target_prob_xgb tune the XGB target probability model
TODO: actually document this rather than just copy paste
}
