% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_play.R
\name{animate_play}
\alias{animate_play}
\title{Plot a play}
\usage{
animate_play(
  game_id = 2018090600,
  play_id = 75,
  plays = import_plays(),
  games = import_games(),
  positions = import_positions(),
  week = NULL,
  tracking = NULL,
  team_colors = FALSE,
  yardmin = NULL,
  yardmax = NULL,
  field_color = "#b3e3d6",
  line_color = "black",
  sideline_color = "white",
  endzone_color = NULL,
  buffer = NULL,
  nearest_defender = FALSE,
  target_probability = FALSE,
  clip = TRUE,
  at = "end_routes",
  init_cnd = dplyr::quos(.data$frame_id == 1L),
  save = TRUE,
  dir = get_bdb_dir_figs(),
  ext = "gif",
  filename = sprintf("\%s-\%s.\%s", game_id, play_id, ext),
  path = file.path(dir, filename),
  subtitle = NULL,
  end_pause = 1,
  height = 600,
  width = 600,
  fps = 10,
  ...
)
}
\arguments{
\item{game_id}{\code{game_id}}

\item{play_id}{\code{play_id}}
}
\description{
Plot a play
}
\examples{
\dontrun{
plot_play()
}
}
