% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{catch_prob_diagnostic_plots}
\alias{catch_prob_diagnostic_plots}
\title{catch_prob_diagnostic_plots make catch prob model diagnostic plots}
\usage{
catch_prob_diagnostic_plots(train, test, xgb_model, mod = "")
}
\arguments{
\item{train}{the training set}

\item{test}{the test set}

\item{xgb_model}{the xgboost model}

\item{mod}{'a' for arrival, 't' for throw}
}
\value{
A success string
}
\description{
catch_prob_diagnostic_plots make catch prob model diagnostic plots
}
