% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_distance.R
\name{compute_min_distances}
\alias{compute_min_distances}
\title{Compute min distances between players.}
\usage{
compute_min_distances(o, d, one_pass = TRUE)
}
\arguments{
\item{o, d}{A data.frame with 3 columns: \code{nfl_id}, \code{x}, \code{y}.}

\item{one_pass}{A boolean indicating whether to pass through defenders just once (for a one-to-one assignment), probably leaving some defenders without a matched offensive player. If \code{FALSE}, two passes are done (to assigned double coverers.}
}
\value{
A tibble with 6 columns: \code{nfl_id_o}, \code{nfl_id_d}, \code{x_o}, \code{x_d}, \code{y_o}, \code{y_d}.
}
\description{
Compute "optimal" nearest assignments of defensive players to individual offensive and defensive players (not including QB) in a single frame using the Hungarian method.
}
