% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_positions}
\alias{import_positions}
\title{Import position categorization data}
\usage{
import_positions(dir = .get_dir())
}
\arguments{
\item{dir}{Directory where file is stored, which is dependent on whether
the code is run locally or on Kaggle.
Best to change with \code{options(bdb2021.dir = <dir>)} so that other \code{import_}
functions have the same change.}
}
\description{
Useful for \code{side} column (either \code{"O"} or \code{"D"})
}
\examples{
\dontrun{
import_positions()
}
}
\seealso{
\url{https://github.com/leesharpe/nfldata/blob/master/data/positions.csv}
}
