% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_positions}
\alias{import_positions}
\title{Import position categorization data}
\usage{
import_positions(dir = .get_dir())
}
\description{
Useful for \code{side} column (either \code{"O"} or \code{"D"})
}
\examples{
\dontrun{
import_positions()
}
}
\seealso{
\url{https://github.com/leesharpe/nfldata/blob/master/data/positions.csv}
}
