% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_play}
\alias{plot_play}
\title{Plot a play}
\usage{
plot_play(
  game_id = 2018090600,
  play_id = 75,
  plays = import_plays(),
  games = import_games(),
  positions = import_positions(),
  week = NULL,
  tracking = NULL,
  team_colors = FALSE,
  yardmin = NULL,
  yardmax = NULL,
  field_color = "#b3e3d6",
  line_color = "black",
  sideline_color = "white",
  endzone_color = NULL,
  buffer = NULL,
  at = "end_routes",
  save = TRUE,
  dir = "inst",
  filename = sprintf("\%s-\%s.png", game_id, play_id),
  path = file.path(dir, filename),
  width = 10,
  height = 10,
  ...
)
}
\arguments{
\item{game_id}{\code{game_id}}

\item{play_id}{\code{play_id}}
}
\description{
Plot a play
}
\examples{
\dontrun{
plot_play()
}

}
