% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{run_catch_prob_tuning_pipeline}
\alias{run_catch_prob_tuning_pipeline}
\title{run_catch_prob_tuning_pipeline runs the catch prob tuning pipeline}
\usage{
run_catch_prob_tuning_pipeline()
}
\value{
data.frame: normalized positional data
}
\description{
run_catch_prob_tuning_pipeline runs the catch prob tuning pipeline
}
