% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{run_catch_prob_tuning_pipelines}
\alias{run_catch_prob_tuning_pipelines}
\title{run_catch_prob_tuning_pipeline runs the catch prob tuning pipeline}
\usage{
run_catch_prob_tuning_pipelines(overnightmode = TRUE, overnightcores = 4)
}
\arguments{
\item{overnightmode}{boolean whether or not to run in overnightmode}

\item{overnightcores}{integer how many cores to run on}
}
\value{
NULL (invisible)
}
\description{
run_catch_prob_tuning_pipeline runs the catch prob tuning pipeline
}
