% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{stepwise_catch_prob_predict}
\alias{stepwise_catch_prob_predict}
\title{stepwise_catch_prob_predict Make calibrated predictions from xgboost + Platt scaling}
\usage{
stepwise_catch_prob_predict(data, xgb_model)
}
\arguments{
\item{data}{a data frame}

\item{xgb_model}{the xgboost (parsnip) model}
}
\value{
a vector of predicted probabilities
}
\description{
stepwise_catch_prob_predict Make calibrated predictions from xgboost + Platt scaling
}
